tasks/testall -> tasks/try-each, generalized to work for more than just test - Remove tasks/test and just use tasks/try-each, for a single scenario - Inject dependencyManagerAdapter & speed up some tests - Appease JSCS - Add note about some tests using a dependency manager for real - Use ember try:testall for "client" test for a smoke test - Remove dead code
