+++ lib/commands/testall.js
@@ -13,14 +13,14 @@ module.exports = {
 
     var config = require('../utils/config')({ project: this.project });
 
-    var TestallTask = require('../tasks/testall');
+    var TryEach = require('../tasks/try-each');
 
-    var testallTask = new TestallTask({
+    var tryEachTask = new TryEach({
       ui: this.ui,
       project: this.project,
       config: config
     });
 
-    return testallTask.run(commandOptions);
+    return tryEachTask.run(config.scenarios, commandOptions);
   }
 };
