+++ lib/commands/try.js
@@ -49,14 +49,15 @@ module.exports = {
                       'specified in the ember-try.js config.');
     }
 
-    var TryTask = require('../tasks/try');
-    var tryTask = new TryTask({
+    var TryEachTask = require('../tasks/try-each');
+    var tryEachTask = new TryEachTask({
       ui: this.ui,
       project: this.project,
-      config: config
+      config: config,
+      commandArgs: commandArgs
     });
 
-    return tryTask.run(scenario, commandArgs, commandOptions);
+    return tryEachTask.run([scenario], commandOptions);
   }
 };
