+++ lib/tasks/try-each.js
@@ -8,12 +8,22 @@ var runCommand      = require('./../utils/run-command');
 var ResultSummary   = require('./../utils/result-summary');
 
 module.exports = CoreObject.extend({
-  run: function(options) {
+  run: function(scenarios, options) {
     var task = this;
-    var scenarios = this.config.scenarios;
-    task.ScenarioManager = new ScenarioManager({ui: task.ui, project: task.project});
+
+    task.ScenarioManager = new ScenarioManager({ui: task.ui,
+                                                project: task.project,
+                                                dependencyManagerAdapter: this.dependencyManagerAdapter});
+
+    process.on('SIGINT', function() {
+      task.ui.writeLine('\nGracefully shutting down from SIGINT (Ctrl-C)');
+      return task.ScenarioManager.cleanup().then(function() {
+        task._exit();
+      });
+    });
+
     return task.ScenarioManager.setup().then(function() {
-      return mapSeries(scenarios, task._runTestsForThisScenario, task);
+      return mapSeries(scenarios, task._runCommandForThisScenario, task);
     }).then(function(results) {
       return task._optionallyCleanup(options).then(function() {
         task._printResults(results);
@@ -29,7 +39,7 @@ module.exports = CoreObject.extend({
     });
   },
 
-  _runTestsForThisScenario: function(scenario) {
+  _runCommandForThisScenario: function(scenario) {
     var task = this;
     return task.ScenarioManager.changeTo(scenario)
       .then(function(scenarioDependencyState) {
@@ -37,16 +47,16 @@ module.exports = CoreObject.extend({
           scenario: scenario.name,
           dependencyState: scenarioDependencyState
         };
-        return task._runTests().then(function(result) {
+        return task._runCommand().then(function(result) {
           runResults.result = result;
           return RSVP.resolve(runResults);
         });
       });
   },
 
-  _runTests: function() {
+  _runCommand: function() {
     var task = this;
-    return runCommand(task.project.root, ['test']);
+    return runCommand(task.project.root, task.commandArgs || ['test']);
   },
 
   _printResults: function(results) {
