+++ lib/utils/bower-adapter.js
@@ -38,14 +38,6 @@ module.exports = CoreObject.extend({
       return RSVP.Promise.resolve(currentDeps);
     });
   },
-  _findCurrentVersionOf: function(packageName) {
-    var filename = path.join(this.cwd, 'bower_components', packageName, 'bower.json');
-    if (fs.existsSync(filename)) {
-      return JSON.parse(fs.readFileSync(filename)).version;
-    } else {
-      throw 'File ' + filename + ' does not exist';
-    }
-  },
   cleanup: function() {
     var adapter = this;
     return adapter._restoreOriginalBowerFile().then(function() {
@@ -56,6 +48,14 @@ module.exports = CoreObject.extend({
     .then(function() {
       return adapter._install();
     });
+  },
+  _findCurrentVersionOf: function(packageName) {
+    var filename = path.join(this.cwd, 'bower_components', packageName, 'bower.json');
+    if (fs.existsSync(filename)) {
+      return JSON.parse(fs.readFileSync(filename)).version;
+    } else {
+      throw 'File ' + filename + ' does not exist';
+    }
   },
   _install: function() {
     var adapter = this;
